Updates tests to parse UCS2 names of test isos.
The test isos have supplementary volume descriptors! The Joliet versions
of many of these wacky names are interesting. Many of them are truncated
(to 128 bytes, in this case) since UCS2 is twice as wide as UTF-8. Many
of the Windows-style forbidden characters are replaced with underscores.

Here, the tests are written to rewrite the test data the way the image
creator likely did: truncates the strings (and removes whitespace from
the truncated version) and replaces forbidden characters with UCS2
underscores.